fix: createOrder 서비스 로직 수정

+++ src/main/java/com/programmers/cafekiosk/dto/ItemResponse.java
@@ -5,9 +5,9 @@ import com.programmers.cafekiosk.entity.ItemType;
 
 import java.time.LocalDateTime;
 
-public record ItemResponse(Long id, ItemType type, ItemStatus status, String name, Integer price, String description, LocalDateTime createdAt, LocalDateTime updatedAt) {
+public record ItemResponse(Long id, ItemType type, ItemStatus status, String name, Long price, String description, LocalDateTime createdAt, LocalDateTime updatedAt) {
 
-    public static ItemResponse from(Long id, ItemType type, ItemStatus status, String name, Integer price, String description, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public static ItemResponse from(Long id, ItemType type, ItemStatus status, String name, Long price, String description, LocalDateTime createdAt, LocalDateTime updatedAt) {
         return new ItemResponse(id, type, status, name, price, description, createdAt, updatedAt);
     }
 }
