+++ src/main/java/com/programmers/cafekiosk/dto/UpdateItemRequest.java
@@ -3,5 +3,5 @@ package com.programmers.cafekiosk.dto;
 import com.programmers.cafekiosk.entity.ItemStatus;
 import com.programmers.cafekiosk.entity.ItemType;
 
-public record UpdateItemRequest(ItemType type, ItemStatus status, String name, Integer price, String description) {
+public record UpdateItemRequest(ItemType type, ItemStatus status, String name, Long price, String description) {
 }
